+++ azext_iot/central/services/device.py
@@ -7,12 +7,14 @@
 
 from typing import List, Union
 import requests
+from azext_iot.common.auth import get_aad_token
 
 from knack.util import CLIError
 from knack.log import get_logger
 
 from azext_iot.constants import CENTRAL_ENDPOINT
 from azext_iot.central.services import _utility
+from azext_iot.central.models.devicetwin import DeviceTwin
 from azext_iot.central.models.preview import DevicePreview
 from azext_iot.central.models.v1 import DeviceV1
 from azext_iot.central.models.v1_1_preview import DeviceV1_1_preview
@@ -497,6 +499,47 @@ def get_component_command_history(
     return _utility.try_extract_result(response)
 
 
+def get_device_twin(
+    cmd,
+    app_id: str,
+    device_id: str,
+    token: str,
+    central_dns_suffix=CENTRAL_ENDPOINT,
+) -> DeviceTwin:
+    """
+    Get device twin given a device id
+
+    Args:
+        cmd: command passed into az
+        device_id: unique case-sensitive device id,
+        app_id: name of app (used for forming request URL)
+        token: (OPTIONAL) authorization token to fetch device details from IoTC.
+            MUST INCLUDE type (e.g. 'SharedAccessToken ...', 'Bearer ...')
+        central_dns_suffix: {centralDnsSuffixInPath} as found in docs
+
+    Returns:
+        twin: dict
+    """
+
+    if not token:
+        aad_token = get_aad_token(cmd, resource="https://apps.azureiotcentral.com")[
+            "accessToken"
+        ]
+        token = "Bearer {}".format(aad_token)
+
+    url = f"https://{app_id}.{central_dns_suffix}/system/iothub/devices/{device_id}/get-twin?extendedInfo=true"
+    headers = _utility.get_headers(token, cmd)
+
+    # Construct parameters
+
+    response = requests.get(
+        url,
+        headers=headers,
+        verify=not should_disable_connection_verify(),
+    )
+    return DeviceTwin(_utility.try_extract_result(response))
+
+
 def run_manual_failover(
     cmd,
     app_id: str,
@@ -572,4 +615,37 @@ def run_manual_failback(
     )
     _utility.log_response_debug(response=response, logger=logger)
 
+    return _utility.try_extract_result(response)
+
+
+def purge_c2d_messages(
+    cmd,
+    app_id: str,
+    device_id: str,
+    token: str,
+    central_dns_suffix=CENTRAL_ENDPOINT,
+) :
+    """
+    Purges cloud to device (C2D) message queue for the specified device.
+
+    Args:
+        cmd: command passed into az
+        app_id: name of app (used for forming request URL)
+        device_id: unique case-sensitive device id,
+        token: (OPTIONAL) authorization token to fetch device details from IoTC.
+            MUST INCLUDE type (e.g. 'SharedAccessToken ...', 'Bearer ...')
+        central_dns_suffix: {centralDnsSuffixInPath} as found in docs
+
+    Returns:
+        {
+            message: 'Cloud to device (C2D) message queue purged for device {device_id}.\\n
+            Total messages purged: {totalMessagesPurged}.'
+        } on success
+        Raises error on failure
+    """
+    url = "https://{}.{}/{}/{}/c2d".format(
+        app_id, central_dns_suffix, "system/iothub/devices", device_id
+    )
+    headers = _utility.get_headers(token, cmd)
+    response = requests.delete(url, headers=headers)
     return _utility.try_extract_result(response)