+++ HISTORY.rst
@@ -3,10 +3,28 @@
 Release History
 ===============
 
+0.12.2
++++++++++++++++
+
+**IoT Central updates**
+
+* Added `az iot central device c2d-message purge` to purge cloud-to-device message queue
+
+
+0.12.1
++++++++++++++++
+
+**IoT DPS updates**
+
+* Resolves issue where usage of `--login` with connection string still required `az login`.
+
+
 0.12.0
 +++++++++++++++
 
 **IoT Central updates**
+
+* Fixed iot hub token leak for device twin show
 
 * Adds new preview commands (v1.1-preview)
 
@@ -14,7 +32,7 @@ Release History
   - Destination (az iot central export destination)
   - Export (az iot central export)
 
-** General Updates **
+**General Updates**
 
 * The IoT extension officially supports Python 3.10.